+++ realm/realm-library/src/main/cpp/io_realm_RealmQuery.cpp
@@ -40,7 +40,8 @@ JNIEXPORT jstring JNICALL Java_io_realm_RealmQuery_nativeSerializeQuery(JNIEnv* 
     return to_jstring(env, "");
 }
 
-JNIEXPORT jlong JNICALL Java_io_realm_RealmQuery_nativeSubscribe(JNIEnv* env, jclass, jlong shared_realm_ptr, jstring j_name, jlong table_query_ptr, jlong descriptor_ptr)
+JNIEXPORT jlong JNICALL Java_io_realm_RealmQuery_nativeSubscribe(JNIEnv* env, jclass, jlong shared_realm_ptr,
+        jstring j_name, jlong table_query_ptr, jlong descriptor_ptr, REALM_UNUSED jlong time_to_live_ms, REALM_UNUSED jboolean update)
 {
     TR_ENTER()
     try {
@@ -50,7 +51,7 @@ JNIEXPORT jlong JNICALL Java_io_realm_RealmQuery_nativeSubscribe(JNIEnv* env, jc
         auto descriptor = reinterpret_cast<DescriptorOrdering*>(descriptor_ptr);
         Results r(realm, *query, *descriptor);
 #if REALM_ENABLE_SYNC
-        RowExpr row = partial_sync::subscribe_blocking(r, name);
+        RowExpr row = partial_sync::subscribe_blocking(r, name, util::Optional<int64_t>(time_to_live_ms), update);
         return to_jlong_or_not_found(row.get_index());
 #endif
     }